feat(05-Callbacks): format #2 starting capsOnly
