+++ 05-Callbacks-Iterators/data/startingCapsOnly.data.js
@@ -1,0 +1,33 @@
+export const uppers = [
+  "A",
+  "B",
+  "C",
+  "D",
+  "E",
+  "F",
+  "G",
+  "H",
+  "I",
+  "J",
+  "K",
+  "L",
+  "M",
+  "N",
+  "O",
+  "P",
+  "Q",
+  "R",
+  "S",
+  "T",
+  "U",
+  "V",
+  "W",
+  "X",
+  "Y",
+  "Z",
+];
+
+export const startCapWords = ["Microsoft", "Tokyo", "Charlie"];
+export const noCapWords = ["party", "car", "banana"];
+export const capsNotAtStart = ["whOOps", "lION", "monkeY"];
+export const mixedWords = [...startCapWords, ...noCapWords, ...capsNotAtStart];
